HashMap
Created the HashMap function with a Set and Get function for setting a list item and getting a value from the key. I need to continue with the project and add the additional functionality.

Also waiting for clarification from Odin Discord regarding treatment of setting the key value pairs, i.e. set additional keys with diff values or replace existing key with new value.